Update .gitignore
I need this file updated in the dev branch before I pull any new changes from dev into my working branch. Otherwise, when I switch to the dev branch, it tries to add the node_modules folder with thousands of files in it to my local copy and causes VSCode to hang forever.